Separation of allocation of parsing objects in 2 memory arenas and combined parameters and vars of methods.
This was 2 changes in one so let's break it:

-- Allocation of parsing objects --
We have 2 passes in type-checking (The first just installs declarations,
the second does full type-checking).
Parsing objects can be split in 2 categories. Those that are needed for the
second pass and those that we don't. For example, from a TypeDeclaration
we only need its methods Buf. The rest can be thrown as what we need from the
rest is saved in a Type.
So, these 2 categories are allocated in separate arenas. The one is thrown
after type-checking pass 1 (having saved what we need). To do that effectively,
we need to save local variables of methods from pass 1 (since LocalDeclarations
are thrown away). Which resulted in combining params and vars into a common
hash table holding "locals". That as a result can now catch some more errors.