Foundation for the generation of vtables.
There's a bunch of work. First of all, to generate vtables, it made
sense to do it when computing offsets since anyway we want to
go through all classes there.

Previously, we went through types in whatever order.
Now, we separate the types in inheritance trees. For example, in this:

class A { }
class B extends A { }
class C extends A { }
class D {}
class E extends D {}

we want this

  A    D
 / \   |
B   C  E

It's easy to make those by keeping in a buffer all the types that don't
have a parent. Those are the roots of the trees (but we now lose the
ability to easily check for cyclic inheritance).
Also, we keep some more info for each type - its children. Children
are more difficult to keep since we can have more than one. We keep
them in a fixed array, of 32 children. This is not that limiting but
in the future we want to make a better expandable vector, that is
not dependent in malloc() et al (i.e. not Buf). Anyway, these
arrays are allocated in MEM::CHILDREN, which is freed completely
after the vtable computation.

Now, we go from the top of the trees to the leaves and the reason
that this is better is because first of all now we don't need
to save info like fields_end and methods_end in the types.
Previously, when we went in whatever order, we had to e.g.
keep it because in the above example, we might visit A, then
B, then D, then C. When we went to D, we moved in a different
inheritance tree and we lose the info, so we had to save it.

Now, what we also do is we keep a hash table that goes down the leaves.
This hash table achieves multiple things at once. Its main purpose
is to help in the generation of vtables. Let's identify the problem.

Consider the case where we didn't have overriding. Then, the vtable
of a child is the virtual table of the parent plus maybe some
more methods that it defines. In that case, the "vtable" could
be just a string that the child just augments and prints. For example:

class A {
  int foo() { }
}
class B {
  int bar() { }
}

Parent has the string "A__foo()" (sth like that is what we will need in LLVM
IR generation) and prints it then the child just augments this
string with B__bar and prints it.

The problem with overriding is that we might need to replace an arbitrary
method in the parent vtable if it's overriden. Like if the parent is:
A__foo
A__bar
A__test

and the B overrides bar, we need to change the middle entry.

This fundamental problem means that we need a more sophisticated DS of some sort.
This data structure should have:
- Quick insertion
- Quick lookup
- Quick replacement
- Should respect order

So we used a serialized hash table, but a kind of specialized one. First of
all, for every method in the vtable, we need well... the Method and its
enclosing type (so that we can print e.g. A__foo()). That's the VMethod.
These are the "values" and we save them in a buffer (i.e. Buf).

We also save the "keys", the ids of the methods, in the hash table.
The hash table is separate-chaining, which consists of bucket lists.
Every bucket has two parallel (compile-time constant size, C-array) buffers.
One for the ids, for quick searching, and one for the indexes.
An index for a specific id is the index of its
respective VMethod (pair) in the serialized buffer.

That actually works very well. We have O(1) insertion (keeping the last
bucket we inserted in every bucket list). The lookup is a little
slower because we have to search the whole list but still it's fast
(given also that the buckets are allocated in the same mostly contiguous
arena, MEM::VTABLE).

We also have quick replacement since note that for each method,
we only need to change its value, not the key, i.e. we don't need
to touch the hash table only the serialized buffer. So for example,
if we want to change A__foo to B__foo, we search for "foo" in the
hash table and we find the index in the buffer. We give a pointer
to this element to the user. The user can now change the contents
of the respective VMethod.

Finally of course, because we insert in a serial buffer, we
preserve the order.

Things get more complicated as we go down the tree. Consider this:

    A
   / \
  B   D
 /
C

We want to give to B and D the same VTable. But the problem is, for all
this to be effective, when we go into B, it changes
the VTable. If we don't copy it on the way down, and because we go
in DFS, when we return back to A and go to D, the VTable will have been
changed by B (note that if we go in BFS, we lose some initial
benefits, i.e. that we don't need to save info because we go
continuously to children).

So, what we basically need is some form of immutable data structure.
Now, we did sort of this thing: Note that in the hash table per se,
we only insert and we _never_ replace or remove. The important
thing here is that in the hash table part of the VTable we don't
change randomly. We only insert sequentially (in the buffers,
not in general).
It's only the serial buffer that we change random elements.
And so what we do going down is we keep copies of the buffer,
so when we return back, we have a fresh new deep_copy to give
to D and the hash table stays as it is.

Now, there are a couple of caveats. First of all, we only
need to make a copy if a type has more than one children,
that's why there's some boilerplate to that.

Second, as we said, the hash table is only incremented
serially, but still, we need to do some resetting when
going back. We need to keep in mind 3 things:
- Save and restore last bucket and number of elements
  for each bucket list. This is important because
  if the bucket is like below after some A
  and after one of its children B:

-------
|     | <-- after B
|     |
|     | <-- after A
|     |
-------

We want to restore it back to "after A" state
when going back. We do that by restoring the number of
elements, which controls where do we store in the last
bucket and we also restore the last bucket itself.

- Make note that the children might possibly have
  allocated more buckets which we don't want to waste.
  For this reason, if a bucket has a "next"
  bucket already, we use it (and overwrite it)
  we don't allocated a new one.

- Finally, when searching, we want to do
  it in strict limits of `nelems` because if
  not (if for example say, search all buckets
  in the bucket list until either we find it
  or we go to NULL), then there may be
  left-overs that we come across
  and seem like we found the element
  when we didn't. Consider this:

class A {
  foo()
}
class B extends A {
  bar()
}
class C extends A {
  bar()
}

In a bucket, going down to A, we may insert
bar(). Going back to A and down to C,

A bucket might look like that
-------
|     |
| bar |
|     | <-- nelems
|     |
-------

Going into to C, we're going to search if bar()
is in the VTable (to check if if
C__bar() is a new method, i.e. was not in the parent
and so in the VTable or a potential overriding
one, whose enclosing member we will replace in the
VTable. More on that later).

If we don't take `nelems` into consideration, searching
the bucket, we'll come across "bar", but "bar"
is a left-over and we shouldn't consider that
"bar" is in the table.

Finally, note that by using all those, we can
check easily using the table if a method
overrides another (validly) or not:
- Search the table.
- If it's not there,
  we don't override, we have a new
  offset and entry in the vtable
  and we also insert it.
- If it is, replace the enclosing
  to the new enclosing (i.e. A__foo -> B__foo).

As a last note, a VTable allocates stuff
in MEM::VTABLE. We deallocate it after
we process a tree since obviously trees
don't share any info.
And when all the trees end, we free
all MEM::CHILDREN.