Cleaning-up LLVM IR generation: Part 1
There are a lot of not good parts of this procedure. A lot of functions
that are too specific, the fact that llvalue_t does not have a type
and we have to pass separately etc.

In this first part we clean up the construction of llvalue_t and
we add a type to it, although still not used.